je fais tarpin de truc

+++ Assets/Matteo/Script/Item.cs
@@ -3,16 +3,16 @@ using UnityEngine;
 
 public class Item
 {
-    public static float GetCost(ItemType type)
+    public static int GetCost(ItemType type)
     {
         switch (type)
         {
             case ItemType.Arme:
-                return 20f * Mathf.Pow(1.5f, GameManager.GetArmeLevel());
+                return (int)(20f * Mathf.Pow(1.5f, GameManager.GetArmeLevel()));
             case ItemType.Chaussure:
-                return 15f * Mathf.Pow(1.5f, GameManager.GetChaussureLevel());
+                return (int)(15f * Mathf.Pow(1.5f, GameManager.GetChaussureLevel()));
             case ItemType.Cravate:
-                return 10f * Mathf.Pow(1.5f, GameManager.GetCravateLevel());
+                return (int)(10f * Mathf.Pow(1.5f, GameManager.GetCravateLevel()));
             default:
                 return -1;
         }
